+++ src/index.ts
@@ -2,6 +2,8 @@
  * Combined index of all the modules.
  */
 export * from "./error/coreError";
+export * from "./factories/factoryBase";
+export * from "./factories/networkClientFactory";
 export * from "./helpers/errorHelper";
 export * from "./helpers/jsonHelper";
 export * from "./helpers/numberHelper";